+++ src/main/php/peer/stomp/Destination.class.php
@@ -1,11 +1,12 @@
 <?php namespace peer\stomp;
 
+use lang\Value;
 use util\Objects;
 
 /**
  * STOMP Destination
  */
-class Destination {
+class Destination implements Value {
   protected $name   = null;
   protected $conn   = null;
 
@@ -48,10 +49,25 @@ class Destination {
   }
 
   /**
-   * Retrieve string representation
-   * 
-   * @return string
+   * Compare
+   *
+   * @param  var $value
+   * @return int
    */
+  public function compareTo($value) {
+    if ($value instanceof self && $this->conn === $value->conn) {
+      return strcmp($this->name, $value->name);
+    } else {
+      return 1;
+    }
+  }
+
+  /** @return string */
+  public function hashCode() {
+    return '@'.$this->name;
+  }
+
+  /** @return string */
   public function toString() {
     return $this->name.' -> '.Objects::stringOf($this->conn, '  ');
   }